Include stack head in cycle detection keys
This allows us to make more progress when we reach a certain node in
“different situations” — that is, when we're looking for different
definitions at some point further in the path.  This is represented by
the head of the symbol stack or scope stack being different.  That
doesn't handle situations where the symbol stacks are different later on
in the stacks, but this is hopefully a useful heuristic for things like
instance members in a Python class.